requirements.txt
transformers==4.52.0
python-dotenv
numpy
fastapi
uvicorn
sentencepiece
einops
